requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.29.0
bcrypt==4.1.2
bidict==0.23.1
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.0
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
makefun==1.15.2
Mako==1.3.2
MarkupSafe==2.1.5
pwdlib==0.2.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-multipart==0.0.9
python-socketio==5.11.1
PyYAML==6.0.1
simple-websocket==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0
